move source to corresponding modules
