+++ cfg4j-git/src/main/java/org/cfg4j/source/git/DefaultConfigFilesProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package git;
+package org.cfg4j.source.git;
 
 import java.nio.file.FileSystems;
 import java.nio.file.Path;
