+++ cfg4j-git/src/test/java/org/cfg4j/source/git/AllButFirstTokenPathResolverTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package git;
+package org.cfg4j.source.git;
 
 import static org.mockito.Mockito.when;
